Merge feature/pinvoke-native into develop

+++ TeamDEV.Asl/PInvoke/Structures/ObjectAttributes.cs
@@ -5,14 +5,14 @@ using System.Runtime.InteropServices;
 namespace TeamDEV.Asl.PInvoke.Structures {
     [StructLayout(LayoutKind.Sequential)]
     public struct ObjectAttributes {
-        public UInt32 Length;
+        public uint Length;
         public IntPtr RootDirectory;
         public IntPtr ObjectName;
-        public UInt32 Attributes;
+        public uint Attributes;
         public IntPtr SecurityDescriptor;
         public IntPtr SecurityQOS;
-        public static UInt32 Size {
-            get { return (UInt32) Marshal.SizeOf(typeof(ObjectAttributes)); }
+        public static int Size {
+            get { return Marshal.SizeOf(typeof(ObjectAttributes)); }
         }
     }
 }
